Add more types for payloads.
