+++ api/agentpb/agent.go
@@ -1,40 +1,74 @@
 package agentpb
 
 import (
-	"context"
-
 	"github.com/golang/protobuf/proto"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/metadata"
 )
 
 //go-sumtype:decl isServerMessage_Payload
 //go-sumtype:decl isAgentMessage_Payload
-//go-sumtype:decl ServerMessagePayload
-//go-sumtype:decl AgentMessagePayload
 
 // Workaround for https://github.com/golang/protobuf/issues/261.
 // Useful for helper functions.
+// TODO Refactor code to use
+// ServerRequestPayload, ServerResponsePayload, AgentRequestPayload, AgentResponsePayload instead.
 type (
 	ServerMessagePayload = isServerMessage_Payload
 	AgentMessagePayload  = isAgentMessage_Payload
 )
 
-//go-sumtype:decl RequestPayload
-//go-sumtype:decl ResponsePayload
+//go-sumtype:decl ServerRequestPayload
+//go-sumtype:decl ServerResponsePayload
+//go-sumtype:decl AgentRequestPayload
+//go-sumtype:decl AgentResponsePayload
 
-type RequestPayload interface{ request() }
-type ResponsePayload interface{ response() }
+type ServerRequestPayload interface {
+	server()
+	request()
+}
 
-func (*Ping) request()                {}
-func (*QANCollectRequest) request()   {}
+type ServerResponsePayload interface {
+	server()
+	response()
+}
+
+type AgentRequestPayload interface {
+	agent()
+	request()
+}
+
+type AgentResponsePayload interface {
+	agent()
+	response()
+}
+
+// code below uses the same order as definitions in agent.proto (except ping/pong)
+
+func (*Ping) agent()    {}
+func (*Ping) server()   {}
+func (*Ping) request()  {}
+func (*Pong) agent()    {}
+func (*Pong) server()   {}
+func (*Pong) response() {}
+
+// AgentMessage request payloads
+func (*StateChangedRequest) agent()   {}
 func (*StateChangedRequest) request() {}
-func (*SetStateRequest) request()     {}
+func (*QANCollectRequest) agent()     {}
+func (*QANCollectRequest) request()   {}
 
-func (*Pong) response()                 {}
-func (*QANCollectResponse) response()   {}
+// AgentMessage response payloads
+func (*SetStateResponse) agent()    {}
+func (*SetStateResponse) response() {}
+
+// ServerMessage response payloads
+func (*StateChangedResponse) server()   {}
 func (*StateChangedResponse) response() {}
-func (*SetStateResponse) response()     {}
+func (*QANCollectResponse) server()     {}
+func (*QANCollectResponse) response()   {}
+
+// ServerMessage request payloads
+func (*SetStateRequest) server()  {}
+func (*SetStateRequest) request() {}
 
 //go-sumtype:decl AgentParams
 
@@ -46,75 +80,3 @@ type AgentParams interface {
 
 func (*SetStateRequest_AgentProcess) agentParams() {}
 func (*SetStateRequest_BuiltinAgent) agentParams() {}
-
-const (
-	mdAgentID       = "pmm-agent-id"
-	mdAgentVersion  = "pmm-agent-version"
-	mdAgentNodeID   = "pmm-agent-node-id"
-	mdServerVersion = "pmm-server-version"
-)
-
-// AgentConnectMetadata represents metadata sent by pmm-agent with Connect RPC method.
-type AgentConnectMetadata struct {
-	ID      string
-	Version string
-}
-
-// AgentServerMetadata represents metadata sent by pmm-managed to pmm-agent.
-type AgentServerMetadata struct {
-	AgentRunsOnNodeID string
-	ServerVersion     string
-}
-
-func getValue(md metadata.MD, key string) string {
-	vs := md.Get(key)
-	if len(vs) == 1 {
-		return vs[0]
-	}
-	return ""
-}
-
-// AddAgentConnectMetadata adds metadata to pmm-agent's Connect RPC call.
-// Used by pmm-agent.
-func AddAgentConnectMetadata(ctx context.Context, md *AgentConnectMetadata) context.Context {
-	return metadata.AppendToOutgoingContext(ctx, mdAgentID, md.ID, mdAgentVersion, md.Version)
-}
-
-// GetAgentConnectMetadata returns pmm-agent's metadata.
-// Used by pmm-managed.
-func GetAgentConnectMetadata(ctx context.Context) AgentConnectMetadata {
-	var res AgentConnectMetadata
-	md, ok := metadata.FromIncomingContext(ctx)
-	if ok {
-		res.ID = getValue(md, mdAgentID)
-		res.Version = getValue(md, mdAgentVersion)
-	}
-	return res
-}
-
-// SendAgentServerMetadata sends metadata to pmm-agent.
-// Used by pmm-managed.
-func SendAgentServerMetadata(stream grpc.ServerStream, md *AgentServerMetadata) error {
-	header := metadata.Pairs(
-		mdAgentNodeID, md.AgentRunsOnNodeID,
-		mdServerVersion, md.ServerVersion,
-	)
-	if err := stream.SendHeader(header); err != nil {
-		return err
-	}
-	return nil
-}
-
-// GetAgentServerMetadata receives metadata from pmm-managed.
-// Used by pmm-agent.
-func GetAgentServerMetadata(stream grpc.ClientStream) (AgentServerMetadata, error) {
-	var res AgentServerMetadata
-	md, err := stream.Header()
-	if err != nil {
-		return res, err
-	}
-
-	res.AgentRunsOnNodeID = getValue(md, mdAgentNodeID)
-	res.ServerVersion = getValue(md, mdServerVersion)
-	return res, nil
-}
